requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
scipy>=1.7.0
bitstring>=3.1.9
